# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "post": {
        "operationId": "WorkspacesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a workspace, OWNER membership, and default categories",
        "tags": ["workspaces"]
      },
      "get": {
        "operationId": "WorkspacesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workspaces available to the current user",
        "tags": ["workspaces"]
      }
    },
    "/api/v1/workspaces/{workspaceId}": {
      "get": {
        "operationId": "WorkspacesController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspaces"]
      },
      "patch": {
        "operationId": "WorkspacesController_update",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspaces"]
      },
      "delete": {
        "operationId": "WorkspacesController_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspaces"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/business-profile": {
      "get": {
        "operationId": "BusinessProfileController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["business-profile"]
      },
      "put": {
        "operationId": "BusinessProfileController_put",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["business-profile"]
      },
      "patch": {
        "operationId": "BusinessProfileController_patch",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["business-profile"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "MembersController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-members"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/members/{memberId}": {
      "get": {
        "operationId": "MembersController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-members"]
      },
      "patch": {
        "operationId": "MembersController_update",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-members"]
      },
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-members"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations": {
      "post": {
        "operationId": "InvitationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-invitations"]
      },
      "get": {
        "operationId": "InvitationsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-invitations"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "InvitationsController_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["workspace-invitations"]
      }
    },
    "/api/v1/workspace-invitations/me": {
      "get": {
        "operationId": "AcceptInvitationController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyInvitationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List pending invitations for the authenticated user email",
        "tags": ["workspace-invitations"]
      }
    },
    "/api/v1/workspace-invitations/{invitationId}/accept": {
      "post": {
        "operationId": "AcceptInvitationController_acceptMine",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Workspace membership created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept an invitation belonging to the authenticated user",
        "tags": ["workspace-invitations"]
      }
    },
    "/api/v1/workspace-invitations/{invitationId}/reject": {
      "post": {
        "operationId": "AcceptInvitationController_rejectMine",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyInvitationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject an invitation belonging to the authenticated user",
        "tags": ["workspace-invitations"]
      }
    },
    "/api/v1/workspace-invitations/accept": {
      "post": {
        "operationId": "AcceptInvitationController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept an invitation using its one-time token",
        "tags": ["workspace-invitations"]
      }
    },
...
  },
  "info": {
    "title": "Finance Platform API",
    "description": "MVP REST API for authentication, multi-tenant workspaces, financial operations, budgets, dashboard, and reports.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mi negocio"
          },
          "type": {
            "type": "string",
            "enum": ["PERSONAL", "BUSINESS"]
          },
          "currencyCode": {
            "type": "string",
            "default": "PEN"
          },
          "countryCode": {
            "type": "string",
            "default": "PE"
          },
          "timezone": {
            "type": "string",
            "default": "America/Lima"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["name", "type"]
      },
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mi negocio"
          },
          "currencyCode": {
            "type": "string",
            "default": "PEN"
          },
          "countryCode": {
            "type": "string",
            "default": "PE"
          },
          "timezone": {
            "type": "string",
            "default": "America/Lima"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "BusinessProfileDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string"
          },
          "taxId": {
            "type": "string",
            "description": "RUC for Peruvian businesses"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "countryCode": {
            "type": "string",
            "default": "PE"
          },
          "department": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        },
        "required": ["legalName"]
      },
      "UpdateBusinessProfileDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string"
          },
          "taxId": {
            "type": "string",
            "description": "RUC for Peruvian businesses"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "countryCode": {
            "type": "string",
            "default": "PE"
          },
          "department": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "addressLine2": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "STAFF", "VIEWER"]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED"]
          }
        }
      },
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["ADMIN", "STAFF", "VIEWER"]
          },
          "expiresInDays": {
            "type": "object",
            "default": 7,
            "minimum": 1,
            "maximum": 30
          }
        },
        "required": ["email", "role"]
      },
      "InvitationWorkspaceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["PERSONAL", "BUSINESS"]
          }
        },
        "required": ["id", "name", "type"]
      },
      "InvitationSenderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["id", "name"]
      },
      "MyInvitationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "workspace": {
            "$ref": "#/components/schemas/InvitationWorkspaceResponseDto"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "ADMIN", "STAFF", "VIEWER"]
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACCEPTED", "REJECTED", "EXPIRED", "CANCELLED"]
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "invitedBy": {
            "$ref": "#/components/schemas/InvitationSenderResponseDto"
          }
        },
        "required": ["id", "workspace", "role", "status", "expiresAt", "createdAt"]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["token"]
      },
...
    }
  }
}